+++ lib/apis/auth_api_manager.dart
@@ -1,14 +1,16 @@
 import 'package:graphql_flutter/graphql_flutter.dart';
-import 'package:mmd_ecommerce_fl_lib/apis/base_api_manager.dart';
+import 'package:mmd_ecommerce_fl_lib/apis/base/base_api_manager.dart';
 import 'package:mmd_ecommerce_fl_lib/common_models/common_models.dart';
 import 'package:mmd_ecommerce_fl_lib/generatedql/auth/graphql_api.dart';
 
 class AuthApiManager extends BaseApiManager {
-  static Future<void> loginApi(
-      String email, String password, Function success, Function fail) async {
+  static Future<void> loginApi(String email, String password,
+      String deviceToken, Function success, Function fail) async {
     var result = await BaseApiManager.mainClient().mutate(MutationOptions(
         documentNode: SignInQuery().document,
-        variables: SignInArguments(email: email, password: password).toJson()));
+        variables: SignInArguments(
+                email: email, password: password, deviceToken: deviceToken)
+            .toJson()));
     if (result.hasException) {
       fail(result);
     } else {