+++ lib/mmd_ecommerce_fl_lib.dart
@@ -1,8 +1,7 @@
-import 'package:mmd_ecommerce_fl_lib/apis/api_keys.dart';
-import 'package:mmd_ecommerce_fl_lib/apis/base_api_manager.dart';
+import 'apis/base/api_keys.dart';
+import 'apis/base/base_api_manager.dart';
 
 class MmdECommerceFlLib {
-
   /// to add the base url
   static void submitBaseUrl(String baseUrl) {
     BaseApiManager.setUrl(baseUrl);
@@ -20,10 +19,18 @@ class MmdECommerceFlLib {
     BaseApiManager.setLanguage(lang.value);
     BaseApiManager.refreshClient();
   }
+
+  /// to enable auto print requests and responses --> use it only with debug mode
+  /// you have to call [enableDebug] after [submitBaseUrl] directly
+  /// if you want to change the setting of debugging later.
+  /// you must call one of these [submitLanguage] or [submitTokeAndTokenType]
+  /// to reset the client ..
+  static void enableDebug(bool isDebuggable) {
+    BaseApiManager.setDebuggable(isDebuggable);
+  }
 }
 
 // Todo --> manage token session with the library ..
 // Todo --> auto refresh token with the library ..
-// Todo --> handel error form api
 // Todo --> use fragments for paging info ..
-// Todo --> add http logger ..
+// Todo --> connect library to directly get code from git hub ..